+++ astropy/stats/tests/test_sigma_clipping.py
@@ -29,29 +29,30 @@ def test_sigma_clip():
         # Amazing, I've got the same combination on my luggage!
         randvar = randn(10000)
 
-        filtered_data = sigma_clip(randvar, 1, 2)
+        filtered_data = sigma_clip(randvar, sigma=1, iters=2)
 
         assert sum(filtered_data.mask) > 0
         assert sum(~filtered_data.mask) < randvar.size
 
-        #this is actually a silly thing to do, because it uses the standard
-        #deviation as the variance, but it tests to make sure these arguments
-        #are actually doing something
-        filtered_data2 = sigma_clip(randvar, 1, 2, varfunc=np.std)
+        # this is actually a silly thing to do, because it uses the
+        # standard deviation as the variance, but it tests to make sure
+        # these arguments are actually doing something
+        filtered_data2 = sigma_clip(randvar, sigma=1, iters=2, stdfunc=np.var)
         assert not np.all(filtered_data.mask == filtered_data2.mask)
 
-        filtered_data3 = sigma_clip(randvar, 1, 2, cenfunc=np.mean)
+        filtered_data3 = sigma_clip(randvar, sigma=1, iters=2,
+                                    cenfunc=np.mean)
         assert not np.all(filtered_data.mask == filtered_data3.mask)
 
         # make sure the iters=None method works at all.
-        filtered_data = sigma_clip(randvar, 3, None)
+        filtered_data = sigma_clip(randvar, sigma=3, iters=None)
 
         # test copying
         assert filtered_data.data[0] == randvar[0]
         filtered_data.data[0] += 1.
         assert filtered_data.data[0] != randvar[0]
 
-        filtered_data = sigma_clip(randvar, 3, None, copy=False)
+        filtered_data = sigma_clip(randvar, sigma=3, iters=None, copy=False)
         assert filtered_data.data[0] == randvar[0]
         filtered_data.data[0] += 1.
         assert filtered_data.data[0] == randvar[0]
@@ -59,9 +60,9 @@ def test_sigma_clip():
         # test axis
         data = np.arange(5) + np.random.normal(0., 0.05, (5, 5)) + \
             np.diag(np.ones(5))
-        filtered_data = sigma_clip(data, axis=0, sig=2.3)
+        filtered_data = sigma_clip(data, axis=0, sigma=2.3)
         assert filtered_data.count() == 20
-        filtered_data = sigma_clip(data, axis=1, sig=2.3)
+        filtered_data = sigma_clip(data, axis=1, sigma=2.3)
         assert filtered_data.count() == 25
 
 
@@ -74,7 +75,7 @@ def test_compare_to_scipy_sigmaclip():
 
         randvar = randn(10000)
 
-        astropyres = sigma_clip(randvar, 3, None, np.mean)
+        astropyres = sigma_clip(randvar, sigma=3, iters=None, cenfunc=np.mean)
         scipyres = stats.sigmaclip(randvar, 3, 3)[0]
 
         assert astropyres.count() == len(scipyres)
@@ -82,7 +83,7 @@ def test_compare_to_scipy_sigmaclip():
 
 
 def test_sigma_clipped_stats():
-    """Test list data with input mask or mask_val (#3268)."""
+    """Test list data with input mask or mask_value (#3268)."""
     # test list data with mask
     data = [0, 1]
     mask = np.array([True, False])
@@ -91,8 +92,8 @@ def test_sigma_clipped_stats():
     assert result[1] == 1.
     assert result[2] == 0.
 
-    # test list data with mask_val
-    result2 = sigma_clipped_stats(data, mask_val=0.)
+    # test list data with mask_value
+    result2 = sigma_clipped_stats(data, mask_value=0.)
     assert result2[0] == 1.
     assert result2[1] == 1.
     assert result2[2] == 0.